+++ src/main/java/com/platform/esgplatformbackend/model/po/CorporationESGPo.java
@@ -12,13 +12,13 @@ import lombok.NoArgsConstructor;
 @Builder
 public class CorporationESGPo {
 
-  private long corporation_id;
+  private int corporation_id;
 
-  private long ESG_total_ranking;
+  private int ESG_total_ranking;
 
-  private long ESG_industry_ranking;
+  private int ESG_industry_ranking;
 
-  private long record_id;
+  private int record_id;
 
   private int esg_id;
 }
